+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+## [Unreleased]
+### Fixed
+ - prevent prefixing calls to variable functions
+
+
+
 ## [0.4.2] - 2020-09-25
 ### Fixed
  - Prevent prefixing special function `onKeyEvent`